Breakout getUploadedString to helper function
